Consolidating codec.go registrations. Moving Acknowledgement result/error to its own file
